% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_density.R
\name{cov_density}
\alias{cov_density}
\title{cov_density}
\usage{
cov_density(
  cov_obj,
  genes,
  add_gene_info = NULL,
  density_color = "#F8766D",
  scale = 2,
  show_row_names = T,
  row_names_side = "left",
  row_names_size = 10,
  row_names_rot = 0,
  gene_info_color_list = NULL,
  gene_info_side = "right",
  anno_param_row_gene = list(show = T, width = 5, border = FALSE, name_size = NULL,
    name_rot = 90, name_side = "bottom"),
  legend_side = "right",
  legend_title_size = NULL,
  legend_text_size = NULL,
  legend_grid_size = grid::unit(4, "mm")
)
}
\arguments{
\item{cov_obj}{A COV object.}

\item{genes}{A vector of genes.}

\item{add_gene_info}{A character vector of `gene_info` names.}

\item{density_color}{A string of color for density chart.}

\item{scale}{A number indicating relative height of density distribution.}

\item{show_row_names}{A logical value indicating whether show row names.}

\item{row_names_side}{The position of row names ("left", "right").}

\item{row_names_size}{The size of row names.}

\item{row_names_rot}{The rotation of row names.}

\item{gene_info_color_list}{A list contains colors for `gene_info` annotation.
e.g. list(source = c("reference" = "red", "novel" = "blue"), length = c("orange", "red"))}

\item{gene_info_side}{The position of `gene_info` annotation ("left" and "right").}

\item{anno_param_row_gene}{A list contains parameters for the gene annotation. These can be any of the following:
"show", "width", "border", "name_size", "name_rot" and "name_side".}

\item{legend_side}{The position of legend ("top", "bottom", "right", "left").}

\item{legend_title_size}{The size of legend title.}

\item{legend_text_size}{The size of legend item labels.}

\item{legend_grid_size}{A \code{\link[grid]{unit}} object for the size of legend grid.}

\item{legend_title}{The text for the legend title.}
}
\description{
cov_density
}
