% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_heatmap.R
\name{cov_heatmap}
\alias{cov_heatmap}
\title{cov_heatmap}
\usage{
cov_heatmap(
  cov_obj,
  add_phen_info = NULL,
  add_gene_info = NULL,
  cov_colors = c("white", "#F8766D"),
  gene_info_color_list = NULL,
  phen_info_color_list = NULL,
  border = T,
  cluster_rows = F,
  clustering_distance_rows = "euclidean",
  clustering_method_rows = "complete",
  row_dend_side = "left",
  row_dend_width = grid::unit(5, "mm"),
  row_sorted = c(),
  show_row_names = F,
  row_names_side = "left",
  row_names_size = 10,
  row_names_rot = 0,
  cluster_columns = F,
  clustering_distance_columns = "euclidean",
  clustering_method_columns = "complete",
  column_dend_side = "top",
  column_dend_height = grid::unit(5, "mm"),
  column_sorted = c(),
  show_column_names = F,
  column_names_side = "bottom",
  column_names_size = 10,
  column_names_rot = 90,
  anno_param_row_phen = list(show = T, width = 5, border = F, name_size = NULL,
    name_rot = 90, name_side = "top"),
  anno_param_column_gene = list(show = T, height = 5, border = FALSE, name_size = NULL,
    name_rot = 0, name_side = "right"),
  anno_param_row_stat = list(show = F, width = 10, border = FALSE, title = "Coverage",
    title_size = 10, title_side = "bottom", title_rot = 0, axis_side = "bottom", axis_at
    = NULL, axis_labels = NULL, axis_labels_size = 8),
  anno_param_column_stat = list(show = F, height = 10, border = FALSE, title =
    "Coverage", title_size = 10, title_side = "left", title_rot = 0, axis_side = "left",
    axis_at = NULL, axis_labels = NULL, axis_labels_size = 8),
  legend_side = "right",
  legend_title = "Coverage",
  legend_title_size = NULL,
  legend_text_size = NULL,
  legend_grid_size = grid::unit(4, "mm"),
  use_raster = NULL
)
}
\arguments{
\item{cov_obj}{A COV object.}

\item{add_phen_info}{A character vector of `phen_info` names.}

\item{add_gene_info}{A character vector of `gene_info` names.}

\item{cov_colors}{A vector of colors or a color mapping function, pass to \code{\link[ComplexHeatmap]{Heatmap}}.}

\item{gene_info_color_list}{A list contains named vector of colors for `gene_info` annotation.
e.g. list(source = c("reference" = "red", "novel" = "blue"), length = c("orange", "red"))}

\item{phen_info_color_list}{A list contains named vector of colors for `phen_info` annotation.
e.g. list(gender = c("Male" = "green", "Female" = "red"), age = c("yellow", "red"))}

\item{border}{A logical value or a string of color indicating whether draw border.}

\item{cluster_rows}{A logical value indicating whether perform clustering on rows.}

\item{clustering_distance_rows}{Method of measuring distance when clustring on rows, pass to \code{\link[stats]{dist}}.}

\item{clustering_method_rows}{Method to perform hierarchical clustering on rows, pass to \code{\link[stats]{hclust}}.}

\item{row_dend_side}{The position of the row dendrogram ("left", "right").}

\item{row_dend_width}{A \code{\link[grid]{unit}} object for the width of the row dendrogram.}

\item{row_sorted}{A vector of sorted row names. It only works when `cluster_rows = F`.}

\item{show_row_names}{A logical value indicating whether show row names.}

\item{row_names_side}{The position of row names ("left", "right").}

\item{row_names_size}{The size of row names.}

\item{row_names_rot}{The rotation of row names.}

\item{cluster_columns}{A logical value indicating whether perform clustering on columns.}

\item{clustering_distance_columns}{Method of measuring distance when clustring on columns, pass to \code{\link[stats]{dist}}.}

\item{clustering_method_columns}{Method to perform hierarchical clustering on columns, pass to \code{\link[stats]{hclust}}.}

\item{column_dend_side}{The position of the column dendrogram ("top", "bottom").}

\item{column_dend_height}{A \code{\link[grid]{unit}} object for the height of the column dendrogram.}

\item{column_sorted}{A vector of sorted column names. It only works when `cluster_columns = F` and `split_block = F`.}

\item{show_column_names}{A logical value indicating whether show column names.}

\item{column_names_side}{The position of column names ("top", "column").}

\item{column_names_size}{The size of column names.}

\item{column_names_rot}{The rotation of column names.}

\item{anno_param_row_phen}{A list contains parameters for the phenotype annotation. These can be any of the following:
"show", "width", "border", "name_size", "name_rot" and "name_side".}

\item{anno_param_column_gene}{A list contains parameters for the gene annotation. These can be any of the following:
"show", "height", "border", "name_size", "name_rot" and "name_side".}

\item{anno_param_row_stat}{A list contains parameters for the stat annotation of rows. These can be any of the following:
"show", "width", "border", "title", "title_size", "title_rot", "title_side", "axis_side", "axis_at", "axis_labels" and "axis_labels_size".}

\item{anno_param_column_stat}{A list contains parameters for the stat annotation of columns. These can be any of the following:
"show", "height", "border", "title", "title_size", "title_rot", "title_side", "axis_side", "axis_at", "axis_labels" and "axis_labels_size".}

\item{legend_side}{The position of legend ("top", "bottom", "right", "left").}

\item{legend_title}{The text for the legend title.}

\item{legend_title_size}{The size of legend title.}

\item{legend_text_size}{The size of legend item labels.}

\item{legend_grid_size}{A \code{\link[grid]{unit}} object for the size of legend grid.}

\item{use_raster}{A logical value indicating whether render the heatmap body as a raster image, pass to \code{\link[ComplexHeatmap]{Heatmap}}}
}
\description{
cov_heatmap
}
