% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_obj.R
\name{get_cov_obj}
\alias{get_cov_obj}
\title{get_cov_obj}
\usage{
get_cov_obj(cov_data, phen_info = NULL, gene_info = NULL)
}
\arguments{
\item{cov_data}{A numeric `matrix` or `data.frame` of the coverage data.
The row names are gene names. The column names are sample names.}

\item{phen_info}{A `data.frame` of phenotype data. The row names are sample name.}

\item{gene_info}{A `data.frame` of gene data. The row names are gene name.

The `cov_data` shouldn't contain gene that is 0 in all sample, and the same for sample.}
}
\description{
Get an object of the COV class.
}
