% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pav_obj.R
\name{get_pav_obj}
\alias{get_pav_obj}
\title{get_pav_obj}
\usage{
get_pav_obj(
  pav_data,
  phen_info = NULL,
  gene_info = NULL,
  add_softcore = T,
  add_private = T,
  softcore_loss_rate = 0.1,
  softcore_p_value = 0.05
)
}
\arguments{
\item{pav_data}{A numeric `matrix` or `data.frame` of PAV data. `0` represent "absence" and `1` represent "presence".
The row names are gene names. The column names are sample names.}

\item{phen_info}{The phenotype data. The row names are sample name.}

\item{gene_info}{The gene data. The row names are gene name.}

\item{add_softcore}{A logical value indicating whether to consider `soft-core` when determining the gene types.}

\item{add_private}{A logical value indicating whether to consider `private` when determining the gene types.}

\item{softcore_loss_rate}{A numeric vector of loss rate.}

\item{softcore_p_value}{A numeric vector of p-value (binomial test).

`add_softcore` and `add_private` are logical values indicating whether to consider "softcore" and "private" when determining the gene types.

If `add_softcore` is `TRUE`, the genes with loss rates not significantly larger than `softcore_loss_rate` will be considered as "softcore" gene. Binomial tests (with a null hypothesis of loss rate < `softcore_loss_rate`) are carried out for each gene. A p-vlaue below `softcore_p_value` means that this gene is lost in a significant proportion and is a distributed gene(loss rate > `softcore_loss_rate`).

If `add_private` is `TRUE`, the genes present in only one sample will be condidered as "private" gene.}
}
\description{
Get an object of the PAV class.
}
