% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pav_cluster.R
\name{pav_cluster}
\alias{pav_cluster}
\title{pav_cluster}
\usage{
pav_cluster(
  pav_obj,
  cluster_distance = "euclidean",
  cluster_method = "complete",
  add_phen_info = NULL,
  phen_info_color_list = NULL,
  sample_name_size = 4,
  mult = 0.1,
  legend_side = "right",
  legend_title_size = NULL,
  legend_text_size = NULL
)
}
\arguments{
\item{pav_obj}{A PAV object.}

\item{cluster_distance}{Method to measure distance, pass to \code{\link[stats]{dist}}.}

\item{cluster_method}{Method to perform hierarchical clustering, pass to \code{\link[stats]{hclust}}.}

\item{add_phen_info}{A character string of `phen_info` names.}

\item{phen_info_color_list}{A list contains named vector of colors for `phen_info` annotation.
e.g. list(gender = c("Male" = "green", "Female" = "red"))}

\item{sample_name_size}{The size of labels.}

\item{mult}{A numebr of multiplicative range expansion factors.}

\item{legend_side}{The position of legend ("top", "bottom", "right", "left").}

\item{legend_title_size}{The size of legend title.}

\item{legend_text_size}{The size of legend item labels.}
}
\description{
Cluster samples based on PAV table.
}
