% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pav_hist.R
\name{pav_hist}
\alias{pav_hist}
\title{pav_hist}
\usage{
pav_hist(
  pav_obj,
  show_ring = T,
  ring_pos_x = 0.5,
  ring_pos_y = 0.6,
  ring_r = 0.3,
  ring_label_size = NA,
  type_colors = NULL,
  x_title = "Sample Number",
  x_title_size = NULL,
  y_title = "Count",
  y_title_size = NULL,
  x_breaks = NULL,
  x_text_size = NULL,
  y_text_size = NULL
)
}
\arguments{
\item{pav_obj}{A PAV object.}

\item{show_ring}{A logical value indicating whether draw ring chart.}

\item{ring_pos_x}{A numeric vector that specifies the x-location of the ring chart, ranging from 0 to 1.}

\item{ring_pos_y}{A numeric vector that specifies the y-location of the ring chart, ranging from 0 to 1.}

\item{ring_r}{The radius of the ring chart, ranging from 0 to 0.5.}

\item{type_colors}{A named vector of colors for types. e.g. c("distributed" = "red")}

\item{x_title}{The text for the x-axis title.}

\item{x_title_size}{The size of x-axis title.}

\item{y_title}{The text for the y-axis title.}

\item{y_title_size}{The size of y-axis title.}

\item{x_breaks}{A numeric vector of break values on the x-axis.}

\item{x_text_size}{The size of tick labels on the x-axis.}

\item{y_text_size}{The size of tick labels on the y-axis.}

\item{ring_labels_size}{The size of labels on the ring chart.}
}
\description{
The genes can be divided into multiple types based on how many samples are shared.
 `pav_hist()` integrate a ring chart and a histogram to showing the number of gene types.
}
