% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pav_pca.R
\name{pav_pca}
\alias{pav_pca}
\title{pav_pca}
\usage{
pav_pca(
  pav_obj,
  center = T,
  scale = F,
  rank = NULL,
  add_phen_info = NULL,
  phen_info_color_list = NULL,
  axis_text_size = NULL,
  axis_title_size = NULL,
  legend_side = "right",
  legend_text_size = NULL,
  legend_title_size = NULL
)
}
\arguments{
\item{pav_obj}{A PAV object.}

\item{center}{A logical value indicating whether the variables should be shifted to be
zero centered, pass to \code{\link[stats]{prcomp}}.}

\item{scale}{A logical value indicating whether the variables should be scaled to have
unit variance before the analysis takes place, pass to \code{\link[stats]{prcomp}}.}

\item{rank}{A number specifying the maximal rank, pass to \code{\link[stats]{prcomp}}.}

\item{add_phen_info}{A character string of `phen_info` names.}

\item{phen_info_color_list}{A list contains named vector of colors for `phen_info` annotation.
e.g. list(gender = c("Male" = "green", "Female" = "red"))}

\item{axis_text_size}{The size of tick labels on axis.}

\item{axis_title_size}{The size of axis title.}

\item{legend_side}{The position of legend ("top", "bottom", "right", "left").}

\item{legend_text_size}{The size of legend item labels.}

\item{legend_title_size}{The size of legend title.}
}
\description{
`pav_pca()` will perform PCA analysis of PAV data using `prcomp()`. The `center`, `scale`, and `rank` will pass to `prcomp()`.
}
