% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pav_stackbar.R
\name{pav_stackbar}
\alias{pav_stackbar}
\title{pav_stackbar}
\usage{
pav_stackbar(
  pav_obj,
  show_relative = F,
  add_phen_info = NULL,
  type_colors = NULL,
  phen_info_color_list = NULL,
  cluster_distance = "euclidean",
  cluster_method = "complete",
  bar_width = 1,
  sample_name_size = 4,
  legend_side = "right",
  legend_title = "Gene",
  legend_title_size = NULL,
  legend_text_size = NULL,
  dend_width = 0.05,
  name_width = 0.1
)
}
\arguments{
\item{pav_obj}{A PAV object.}

\item{show_relative}{A logical value indicating whether show relative value.}

\item{add_phen_info}{A character string of `phen_info` names.}

\item{type_colors}{A named vector of colors for types. e.g. c("distributed" = "red")}

\item{phen_info_color_list}{A list contains named vector of colors for `phen_info` annotation.
e.g. list(gender = c("Male" = "green", "Female" = "red"))}

\item{cluster_distance}{Method to measure distance, pass to \code{\link[stats]{dist}}.}

\item{cluster_method}{Method to perform hierarchical clustering, pass to \code{\link[stats]{hclust}}.}

\item{bar_width}{A numeric vector giving the relative width of bars, ranging from 0 to 1.}

\item{sample_name_size}{The size of sample names.}

\item{legend_side}{The position of legend ("top", "bottom", "right", "left").}

\item{legend_title}{The text for the legend title.}

\item{legend_title_size}{The size of legend title.}

\item{legend_text_size}{The size of legend item labels.}

\item{dend_width}{The relative width of dendrogram, ranging from 0 to 1.}

\item{name_width}{The relative width of sample names, ranging from 0 to 1.}
}
\description{
The composition of genes in all samples can be viewed in `pav_stackbar()`.
The chart consists of a hierarchical cluster tree and a bar plot.
}
