% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phen_plot.R
\name{phen_bar}
\alias{phen_bar}
\title{phen_bar}
\usage{
phen_bar(
  pav_obj,
  phen_name,
  gene_name,
  pav_colors = c("#F8766D", "#00BFC4"),
  bar_width = 0.8,
  x_text_size = NULL,
  x_title_size = NULL,
  y_text_size = NULL,
  y_title_size = NULL,
  legend_side = "top",
  legend_title_size = NULL,
  legend_text_size = NULL
)
}
\arguments{
\item{pav_obj}{A PAV object.}

\item{phen_name}{The name of phenotype.}

\item{gene_name}{The name of gene.
.}

\item{pav_colors}{A vector of colors for presence and absence.}

\item{bar_width}{A numeric vector giving the relative width of bars, ranging from 0 to 1.}

\item{x_text_size}{The size of tick labels on x-axis.}

\item{x_title_size}{The size of x-axis title.}

\item{y_text_size}{The size of tick labels on y-axis.}

\item{y_title_size}{The size of y-axis title.}

\item{legend_side}{The position of legend.}

\item{legend_title_size}{The size of legend title.}

\item{legend_text_size}{The size of legend item labels.}
}
\description{
Draw the presence/absence of a specified gene in a specified phenotype group in a barplot.
}
