% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phen_plot.R
\name{phen_heatmap}
\alias{phen_heatmap}
\title{phen_heatmap}
\usage{
phen_heatmap(
  pav_obj,
  phen_stat_res,
  add_gene_info = NULL,
  p_threshold = 0.01,
  adjust_p = T,
  only_show_significant = T,
  flip = F,
  p_colors = c("#1874CD", "#B0E2FF"),
  na_col = "gray",
  cell_border_color = NA,
  gene_info_color_list = NULL,
  cluster_rows = F,
  clustering_distance_rows = "euclidean",
  clustering_method_rows = "complete",
  row_dend_side = "left",
  row_dend_width = grid::unit(5, "mm"),
  row_sorted = c(),
  show_row_names = F,
  row_names_side = "right",
  row_names_size = 10,
  row_names_rot = 0,
  cluster_columns = F,
  clustering_distance_columns = "euclidean",
  clustering_method_columns = "complete",
  column_dend_side = "top",
  column_dend_height = grid::unit(5, "mm"),
  column_sorted = c(),
  show_column_names = F,
  column_names_side = "bottom",
  column_names_size = 10,
  column_names_rot = 90,
  anno_param_gene = list(show = T, width = 5, border = FALSE, name_size = NULL,
    name_rot = 90, name_side = "bottom"),
  legend_side = "right",
  legend_title = list(pav = "PAV", type = "gene"),
  legend_title_size = NULL,
  legend_text_size = NULL,
  legend_grid_size = grid::unit(4, "mm")
)
}
\arguments{
\item{pav_obj}{A PAV object.}

\item{phen_stat_res}{The result from \code{\link[vPan]{phen_stat}}.}

\item{add_gene_info}{A character vector of `gene_info` names.}

\item{p_threshold}{The threshold of p_value/p_adjusted.}

\item{adjust_p}{A logical value indicating whether adjust p_value.}

\item{only_show_significant}{A logical value indicating whether only show p_value/p_adjusted that satisfies the condition.}

\item{flip}{A logical value indicating whether flip cartesian coordinates.}

\item{p_colors}{A vector of colors or a color mapping function for p_value/p_adjusted, pass to \code{\link[ComplexHeatmap]{Heatmap}}.}

\item{na_col}{A string of color for NA values.}

\item{cell_border_color}{`NA` or a string of color for the border of cells.}

\item{gene_info_color_list}{A list contains named vector of colors for `gene_info` annotation.
e.g. list(source = c("reference" = "red", "novel" = "blue"), length = c("orange", "red"))}

\item{cluster_rows}{A logical value indicating whether perform clustering on rows.}

\item{clustering_distance_rows}{Method of measuring distance when clustring on rows, pass to \code{\link[stats]{dist}}.}

\item{clustering_method_rows}{Method to perform hierarchical clustering on rows, pass to \code{\link[stats]{hclust}}.}

\item{row_dend_side}{The position of the row dendrogram ("left", "right").}

\item{row_dend_width}{A \code{\link[grid]{unit}} object for the width of the row dendrogram.}

\item{row_sorted}{A vector of sorted row names. It only works when `cluster_rows = F`.}

\item{show_row_names}{A logical value indicating whether show row names.}

\item{row_names_side}{The position of row names ("left", "right").}

\item{row_names_size}{The size of row names.}

\item{row_names_rot}{The rotation of row names.}

\item{cluster_columns}{A logical value indicating whether perform clustering on columns.}

\item{clustering_distance_columns}{Method of measuring distance when clustring on columns, pass to \code{\link[stats]{dist}}.}

\item{clustering_method_columns}{Method to perform hierarchical clustering on columns, pass to \code{\link[stats]{hclust}}.}

\item{column_dend_side}{The position of the column dendrogram ("top", "bottom").}

\item{column_dend_height}{A \code{\link[grid]{unit}} object for the height of the column dendrogram.}

\item{column_sorted}{A vector of sorted column names. It only works when `cluster_columns = F` and `split_block = F`.}

\item{show_column_names}{A logical value indicating whether show column names.}

\item{column_names_side}{The position of column names ("top", "column").}

\item{column_names_size}{The size of column names.}

\item{column_names_rot}{The rotation of column names.}

\item{anno_param_gene}{A list contains parameters for the phenotype annotation. These can be any of the following:
"show", "width", "border", "name_size", "name_rot" and "name_side".}

\item{legend_side}{The position of legend ("top", "bottom", "right", "left").}

\item{legend_title}{The text for the legend title.}

\item{legend_title_size}{The size of legend title.}

\item{legend_text_size}{The size of legend item labels.}

\item{legend_grid_size}{A \code{\link[grid]{unit}} object for the size of legend grid.}
}
\description{
Plot the result of phenotypic association in a heatmap.
}
