% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phen_stat.R
\name{phen_stat}
\alias{phen_stat}
\title{phen_stat}
\usage{
phen_stat(
  pav_obj,
  phen_name,
  p_adjust_method = "fdr",
  pallel = F,
  parallel_n = parallel::detectCores() - 1
)
}
\arguments{
\item{pav_obj}{A PAV object.}

\item{phen_name}{The name of phenotype used for calculation.}

\item{p_adjust_method}{The adjustment methods, pass to  \code{\link[stats]{p.adjust}}}

\item{pallel}{A logical value indicating whether to use parallel computing.}

\item{parallel_n}{The number of CPU cores used for parallel computing.}
}
\description{
Perform phenotypic association calculations
}
