% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_multi.R
\name{sim_multi_stat}
\alias{sim_multi_stat}
\title{sim_multi_stat}
\usage{
sim_multi_stat(
  pav_obj,
  phen_name,
  n_simulation = 10,
  parallel = FALSE,
  parallel_n = parallel::detectCores() - 1
)
}
\arguments{
\item{pav_obj}{A PAV object.}

\item{phen_name}{The name of phenotype used for grouping.}

\item{n_simulation}{The number of simulations.}

\item{parallel}{A logical value indicating whether to use parallel computing.}

\item{parallel_n}{The number of CPU cores used for parallel computing.}
}
\description{
Perform genome simulation calculations for groups.
}
