% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_stat.R
\name{sim_stat}
\alias{sim_stat}
\title{sim_stat}
\usage{
sim_stat(
  pav_obj,
  genome_type = c("pan", "core"),
  n_simulation = 10,
  parallel = FALSE,
  parallel_n = parallel::detectCores() - 1
)
}
\arguments{
\item{pav_obj}{A PAV object.}

\item{genome_type}{A vector of genome types. These can be any of the following: "pan", "core" and "private".}

\item{n_simulation}{The number of simulations.}

\item{parallel}{A logical value indicating whether to use parallel computing.}

\item{parallel_n}{The number of CPU cores used for parallel computing.}
}
\description{
Perform genome simulation calculations.
}
